LegacyAlgorithm

+++ src/functions.cpp
@@ -58,19 +58,19 @@ void warning() {
   delay(800);
 }
 
-float AngleBetweenPoints(float start_x, float start_y, float end_x, float end_y) {
-    float dx = end_x - start_x;
-    float dy = end_y - start_y;
+float AngleBetweenPoints(double start_x, double start_y, double end_x, double end_y) {
+    double dx = end_x - start_x;
+    double dy = end_y - start_y;
     return normalize_angle(atan2f(dy, dx));
 }
 
-float DistanceFromLine(float x, float y,
-                       float start_x, float start_y,
-                       float end_x,   float end_y) {
-    float dx = end_x - start_x;
-    float dy = end_y - start_y;
+double DistanceFromLine(double x,   double y,
+                       double start_x, double start_y,
+                       double end_x,   double end_y) {
+    double dx = end_x - start_x;
+    double dy = end_y - start_y;
 
-    float length = sqrtf(dx * dx + dy * dy);
+    double length = sqrtf(dx * dx + dy * dy);
     if (length < 1e-6f) return 0.0f;  // degenerate line (start == end)
 
     // Cross product of line vector and point vector
@@ -78,41 +78,41 @@ float DistanceFromLine(float x, float y,
     return (dy * (x - start_x) - dx * (y - start_y)) / length;
 }
 
-float DistanceBetweenPoints(float x1, float y1, float x2, float y2) {
-    float dx = x2 - x1;
-    float dy = y2 - y1;
+double DistanceBetweenPoints(double x1, double y1, double x2, double y2) {
+    double dx = x2 - x1;
+    double dy = y2 - y1;
     return sqrtf(dx*dx + dy*dy);
 }
 
-bool has_passed_goal(float x, float y, float start_x, float start_y, float end_x, float end_y) {
-    float dx = end_x - start_x;
-    float dy = end_y - start_y;
-    float length_sq = dx*dx + dy*dy;
+bool has_passed_goal(double x, double y, double start_x, double start_y, double end_x, double end_y) {
+    double dx = end_x - start_x;
+    double dy = end_y - start_y;
+    double length_sq = dx*dx + dy*dy;
     
     if (length_sq < 0.001f) return true;
     
-    float rx = x - start_x;
-    float ry = y - start_y;
-    float t = (rx*dx + ry*dy) / length_sq;
+    double rx = x - start_x;
+    double ry = y - start_y;
+    double t = (rx*dx + ry*dy) / length_sq;
     
     return (t >= 1.0f);  // Passed the endpoint
 }
 
 float normalize_angle(float radians) {
-    radians = fmodf(radians, 2.0f * M_PI);
-    if (radians < 0) radians += 2.0f * M_PI;
+    radians = fmod(radians, 2.0 * M_PI);
+    if (radians < 0) radians += 2.0 * M_PI;
     return radians;
 }
 
 float angle_diff(float current, float target) {
-    float diff = fmodf(target - current, 2.0f * M_PI);
+    float diff = fmod(target - current, 2.0 * M_PI);
     if (diff > M_PI)  diff -= 2.0f * M_PI;
     if (diff < -M_PI) diff += 2.0f * M_PI;
     return diff;
 }
 
-int sign_of(float value) {
-    if (value > 0.0f) return  1;
-    if (value < 0.0f) return -1;
+int sign_of(double value) {
+    if (value > 0.0) return  1;
+    if (value < 0.0) return -1;
     return 0;
 }
